"PEP8"

+++ app.py
@@ -4,15 +4,25 @@ import httplib2
 import json
 import requests
 
-from flask import Flask, render_template, request, redirect, jsonify, url_for, make_response, flash
-from flask import session as login_session
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from database_setup import Base, Category, Item
+from database_setup import Base, Category, Item, User
 
 from oauth2client.client import flow_from_clientsecrets
 from oauth2client.client import FlowExchangeError
+
+from flask import (
+    Flask,
+    render_template,
+    request,
+    redirect,
+    jsonify,
+    url_for,
+    make_response,
+    flash,
+    session as login_session
+)
 
 CLIENT_ID = json.loads(
     open('client_secrets.json', 'r').read())['web']['client_id']
@@ -30,7 +40,11 @@ db_session = DBSession()
 # Create anti-forgery state token
 @app.route('/login')
 def show_login():
-    state = ''.join(random.choice(string.ascii_uppercase + string.digits) for _ in range(32))
+    state = ''.join(
+        random.choice(
+            string.ascii_uppercase + string.digits
+        ) for _ in range(32)
+    )
     login_session['state'] = state
     # return "The current session state is %s" % login_session['state']
     return render_template('login.html', STATE=state)
@@ -38,7 +52,6 @@ def show_login():
 
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
-    print("hey")
     # Validate state token
     if request.args.get('state') != login_session['state']:
         response = make_response(json.dumps('Invalid state parameter.'), 401)
@@ -90,8 +103,10 @@ def gconnect():
     stored_access_token = login_session.get('access_token')
     stored_gplus_id = login_session.get('gplus_id')
     if stored_access_token is not None and gplus_id == stored_gplus_id:
-        response = make_response(json.dumps('Current user is already connected.'),
-                                 200)
+        response = make_response(
+            json.dumps('Current user is already connected.'),
+            200
+        )
         response.headers['Content-Type'] = 'application/json'
         return response
 
@@ -106,8 +121,13 @@ def gconnect():
 
     data = answer.json()
 
-    login_session['picture'] = data['picture']
     login_session['email'] = data['email']
+    login_session['picture'] = data['picture']
+
+    user_id = get_user_id(login_session['email'])
+    if not user_id:
+        user_id = create_user()
+    login_session['user_id'] = user_id
 
     output = ''
     output += '<h1>Welcome, '
@@ -116,7 +136,8 @@ def gconnect():
     output += '<img src="'
     output += login_session['picture']
     output += ' " style = "width: 300px; height: 300px;border-radius: ' \
-              '150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+              '150px;-webkit-border-radius:' \
+              ' 150px;-moz-border-radius: 150px;"> '
     flash("you are now logged in as %s" % login_session['email'])
     print("done!")
     return output
@@ -127,13 +148,17 @@ def gdisconnect():
     access_token = login_session.get('access_token')
     if access_token is None:
         print('Access Token is None')
-        response = make_response(json.dumps('Current user not connected.'), 401)
+        response = make_response(
+            json.dumps('Current user not connected.'),
+            401
+        )
         response.headers['Content-Type'] = 'application/json'
         return response
     print('In gdisconnect access token is %s', access_token)
     print('User email is: ')
     print(login_session['email'])
-    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s'\
+          % login_session['access_token']
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
     print('result is ')
@@ -147,9 +172,42 @@ def gdisconnect():
         response.headers['Content-Type'] = 'application/json'
         return response
     else:
-        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        response = make_response(
+            json.dumps(
+                'Failed to revoke token for given user.',
+                400
+            )
+        )
         response.headers['Content-Type'] = 'application/json'
         return response
+
+
+def create_user():
+    """This method create a new user"""
+    new_user = User(
+        email=login_session['email'],
+        picture=login_session['picture']
+    )
+    db_session.add(new_user)
+    db_session.commit()
+    user = db_session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+
+def get_user_info(user_id):
+    """This method takes the user id and return the user object"""
+    user = db_session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def get_user_id(email):
+    """This method takes the user email and return the user id"""
+    try:
+        user = db_session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        """PEP8 gives warning here but is needed."""
+        return None
 
 
 @app.route('/')
@@ -158,8 +216,12 @@ def show_index():
     is_in = 'email' in login_session
     print(is_in)
     categories = db_session.query(Category).all()
-    latest_items = db_session.query(Item).order_by(Item.date.desc()).limit(10).all()
-    return render_template('index.html', categories=categories, latest_items=latest_items, is_in=is_in)
+    latest_items = db_session.query(Item)\
+        .order_by(Item.date.desc()).limit(10).all()
+    return render_template('index.html',
+                           categories=categories,
+                           latest_items=latest_items,
+                           is_in=is_in)
 
 
 @app.route('/catalog/<string:category_name>/items')
@@ -167,14 +229,20 @@ def show_category(category_name):
     is_in = 'email' in login_session
     categories = db_session.query(Category).all()
     category = db_session.query(Category).filter_by(name=category_name).one()
-    items = db_session.query(Item).filter_by(category=category).order_by(Item.date.desc()).all()
-    return render_template('category.html', categories=categories, category=category, items=items, is_in=is_in)
+    items = db_session.query(Item).filter_by(category=category)\
+        .order_by(Item.date.desc()).all()
+    return render_template('category.html',
+                           categories=categories,
+                           category=category,
+                           items=items,
+                           is_in=is_in)
 
 
 @app.route('/catalog/<string:category_name>/<string:item_name>/')
 def show_item(category_name, item_name):
     category = db_session.query(Category).filter_by(name=category_name).one()
-    item = db_session.query(Item).filter_by(name=item_name, category_id=category.id).one()
+    item = db_session.query(Item)\
+        .filter_by(name=item_name, category_id=category.id).one()
     return render_template('item.html', item=item)
 
 
@@ -183,8 +251,10 @@ def create_item():
     if 'email' not in login_session:
         return redirect('/login')
     if request.method == 'POST':
-        new_item = Item(name=request.form['name'], description=request.form['description'],
-                        category_id=request.form['category_id'])
+        new_item = Item(name=request.form['name'],
+                        description=request.form['description'],
+                        category_id=request.form['category_id'],
+                        user_id=login_session['user_id'])
         db_session.add(new_item)
         db_session.commit()
         return redirect(url_for('show_index'))
@@ -198,17 +268,25 @@ def edit_item(item_name):
     if 'email' not in login_session:
         return redirect('/login')
     item_to_be_edited = db_session.query(Item).filter_by(name=item_name).one()
+    if item_to_be_edited.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are" \
+               " not authorized to " \
+               "edit this item. Please create your own item in order to " \
+               "edit.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         item_to_be_edited.name = request.form['name']
         item_to_be_edited.description = request.form['description']
         item_to_be_edited.category_id = request.form['category_id']
         db_session.add(item_to_be_edited)
         db_session.commit()
-        category = db_session.query(Category).filter_by(id=item_to_be_edited.category_id).one()
+        category = db_session.query(Category)\
+            .filter_by(id=item_to_be_edited.category_id).one()
         return redirect(url_for('show_category', category_name=category.name))
     else:
         categories = db_session.query(Category).all()
-        return render_template('edit_item.html', categories=categories, item=item_to_be_edited)
+        return render_template('edit_item.html',
+                               categories=categories,
+                               item=item_to_be_edited)
 
 
 @app.route('/catalog/<string:item_name>/delete', methods=['GET', 'POST'])
@@ -216,10 +294,16 @@ def delete_item(item_name):
     if 'email' not in login_session:
         return redirect('/login')
     item_to_be_deleted = db_session.query(Item).filter_by(name=item_name).one()
+    if item_to_be_deleted.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert(" \
+               "'You are not authorized to delete this item. Please create" \
+               " your own item in order to " \
+               "delete.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         db_session.delete(item_to_be_deleted)
         db_session.commit()
-        category = db_session.query(Category).filter_by(id=item_to_be_deleted.category_id).one()
+        category = db_session.query(Category)\
+            .filter_by(id=item_to_be_deleted.category_id).one()
         return redirect(url_for('show_category', category_name=category.name))
     else:
         return render_template('delete_item.html', item=item_to_be_deleted)
@@ -240,4 +324,4 @@ def catalog_json():
 if __name__ == '__main__':
     app.secret_key = 'super_secret_key'
     app.debug = True
-    app.run(host='0.0.0.0', port=5000)
+    app.run(host='0.0.0.0', port=8080)
